requirements.txt
Flask
Flask-SQLAlchemy
Flask-Admin
Pillow
werkzeug